pyproject.toml
[project]
name = "mnq-position-sizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cx-freeze>=8.3.0",
    "pillow>=11.3.0",
    "pyinstaller>=6.14.1",
    "ttkbootstrap>=1.12.1",
]
